Callbacks implemented
We can now receive callbacks. It looks like only "wheel" is firing. We
need to figure out how to convert the generic event into something
usable and then we can calculate the viewspace properly. After that, we
can move the renderer logic to renderer and attempt to hook up the
actual tiles.