Zoom works with test data!
We can now zoom in and out and have some fake data rendered properly.

The fake data is properly "generated" from the model and is returned via
an iterator that queries the model for the tile at a given xy
coordinate. Unfortunately I couldn't figure out how to remove the mut
from self so we have to do this compute/iterator in two distinct calls.
Maybe there's a proper fix to combine these but I'm ignoring for now.

The renderer now takes in the canvas and context. This means it is
actually rendering now.

The listeners were not really doing anything but holding refcounts. I've
now stored them in a vec.

Now we have to fix the iterator bug dropping a square, get the view
calculation to work for the tile_range() params, and implement dragging.