Implement zoom and pan!
Still super buggy but it works.
Fixed the bug with the TileView.
Hook up TileView to the renderer.
Implement compute() -> tile_range() binding.
Move cursor logic to ViewPort struct and send all cursor related events
to it. If it says the viewport has changed then we'll render the
changes.
This also unifies the events. Or tries to.
We also calculate which tiles are visible to our viewport. This all
relates to the tile_range() logic.
Now panics!() print to the console.