Can now scroll forever!
Fixed that annoying bug where the tiles would disappear if you scrolled
to far. Turns out that applying mod axis to the offset before giving it
to the caller was a mistake! If we did everything relative then it'd be
ok but I guess we don't. We may want to revisit that idea as it does
seem better. It's also kind of concerning that I thought I was already
doing that but wasn't. My logic must be so fucked up.