Clean up errors, trim some fat
Remove a bunch of web-sys feature dependencies that we never used.
Hopefully it speeds up builds and clarifies what we're using in the
code.

Clean up some rust warnings by naming fields _whatever and putting pub
where it needs to go. Longer term we'll have to rearrange some code but
this is a good step for the time being.

Add a screen-saver flag so we can keep the scrolling behavior.