Initial changes to support touch gestures
We can scroll around on a phone now. We need to do a bit of extra work
to support zooming on mobile though. At the moment we only handle 1
point of contact but we'll need to handle at least 2 to get
pinch-to-zoom working.

Also, no clue how to get the god damn canvas to fill up the screen. How
the hell is this so hard?