Rename stuff
Previously we had calcada as the container but it is now more
appropriately named mosaic.

What was mosaic (essentially the interpreter) is now called tessera.
This is neat as a tessera is only a component of a mosaic. That applies!
The interpreter only stores the current state and the mosaic is the full
trace.

Ok, ONE LAST feature request. Shareable colorschemes! Gotta figure out
how to expose all of the params back into the url. Maybe it's easy? :'(